requirements.txt
Django>=4.2
uvicorn>=0.22.0
whitenoise>=6.4.0
psycopg2-binary>=2.9.6
asgiref>=3.8.0
python-dotenv>=1.0.0
